multiple fixes  and tweaks to the test client: it was quickly overloading my 4 core server thanks to an absurd number of printf()+send() calls; the first iteration changed this to using send() only with constant strings, but profiling showed that now the CPU load was entirely in System, i.e. the kernel. The second iteration is shown here: a large buffer is prepared with constant data, which is accessible by every thread for pumping across the line.
Also, the threads are not suspended when the 15 seconds of the test expire, but are told to 'bugger off', i.e. abort. Thanks to using select() instead of polling the socket like crazy, this is easily done with minimal load by the test app during tests.

Further fixes: using the correct (64-bit) types for the CPU affinity mask + corrected the modulo logic as well (shift-left by 999? heh. :-) )