CGI: revert and correct the 'CGI interpreter with spaces in path' issue: issue 77 indicates that you also may want to specify command line arguments with the CGI interpreter like so: '-I interp.exe -opt1 -op2' so we should be able to support '-I '"path to interp.exe" -r -v', however, Windows does NOT correctly process single quotes that way, so you MUST specify [-I "'path to interp.exe' -opt1 -opt2"] and then the single quotes in there will screw things up very badly at the time of CreateProcess, effectively making every run fail immediately. The 'solution' is to monitor the set_option logic and 'transform' any single quotes to double quotes around any incoming CGI interpreter path.
Of course, this new fix/kludge does NOT solve matters when one of those interpreter commandline options also needs double quotes; in that case the only way forward is using a config file, rather than specifying these directly on the command line.

The previous tweak has been reverted because it would break the possibility to pass arguments with the binary in the CGI interpreter option as it would surround the entire thing with double quotes, which is another wrong.